# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: go-proxy-manager admin API
  version: 0.1.0
  summary: 'REST API for the go-proxy-manager control plane: config CRUD, settings, backup/restore, DNS
    sync, Ingress discovery, and admin auth.'
  description: 'Covers the admin control-plane surface only (port 8081 by default), not the public data
    plane that serves proxied traffic. Every non-2xx response uses the uniform `{"error": "..."}` shape
    (see #/components/schemas/Error), except the plain-text/HTML auth endpoints noted individually below.


    Authentication: an admin session cookie (`sessionCookie`) or a scoped API bearer token (`bearerAuth`)
    - either satisfies every operation below unless noted. A session''s access is governed by its role
    (`user` for GET /api/me only; `admin` for everything under /api/); a token''s access is additionally
    governed by its granted scopes, shown per-operation as `x-scope` (`admin` satisfies every scope check;
    a concrete `<plural>:write` implies `<plural>:read`; `*:read`/`*:write` must be granted literally
    - a resource-scoped grant does not satisfy a wildcard requirement). Mutating session requests additionally
    require the `X-CSRF-Token` header (double-submit against the value GET /api/me returns); bearer requests
    never need it.


    HA: on an instance running GPM_HA_ROLE=follower, every write (any POST/PUT/DELETE under /api/) answers
    503 rather than being applied; reads are unaffected. See docs/ha.md.'
  license:
...
servers:
...
security:
- sessionCookie: []
- bearerAuth: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: gpm_session
      description: Admin browser session, issued by /auth/login (OIDC) or /auth/local (bcrypt). Cookie
        is HttpOnly, SameSite=Lax, and named __Host-gpm_session when GPM_COOKIE_SECURE=1 (default). A
        session principal has whatever access its role (admin/user) grants and is unaffected by API-token
        scopes. Mutating requests also need the X-CSRF-Token header.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: gpm_<random>
      description: A scoped API token (config kind api-tokens), created via PUT /api/api-tokens/{name}.
        Every token principal is admin-role by construction; real authorization is per-route scope checking
        (see each operation's x-scope). No CSRF header is used or required.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required:
      - error
      description: Uniform error shape returned by every endpoint on a non-2xx response.
...
paths:
...
  /auth/local:
    post:
      summary: Local username/password login
      description: 'Same-origin only (CSRF belt: no session exists yet for a token-based defense). Rate-limited
        per client IP.'
      tags:
      - Auth
      responses:
        '302':
          description: Redirect to the return path, with the session cookie set.
        '403':
          description: Cross-origin request blocked (failed the same-origin guard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many recent failed attempts from this client IP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Bad credentials, or local login is disabled/not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                return:
                  type: string
                  description: Same-site path to redirect to on success.
...